@@ -4,65 +4,49 @@
   "private": true,
   "scripts": {
     "build": "next build",
+    "db:push": "dotenv drizzle-kit push:mysql",
+    "db:studio": "dotenv drizzle-kit studio",
     "dev": "next dev",
-    "postinstall": "prisma generate",
     "lint": "next lint",
     "start": "next start"
   },
   "dependencies": {
-    "@next-auth/prisma-adapter": "^1.0.5",
-    "@planetscale/database": "^1.7.0",
-    "@prisma/client": "^4.11.0",
-    "@radix-ui/react-collapsible": "^1.0.2",
-    "@radix-ui/react-dialog": "^1.0.3",
-    "@radix-ui/react-dropdown-menu": "^2.0.4",
-    "@radix-ui/react-label": "^2.0.1",
-    "@radix-ui/react-separator": "^1.0.2",
-    "@radix-ui/react-toast": "^1.1.3",
-    "@t3-oss/env-nextjs": "^0.2.1",
-    "@tanstack/react-query": "^4.28.0",
-    "@trpc/client": "^10.18.0",
-    "@trpc/next": "^10.18.0",
-    "@trpc/react-query": "^10.18.0",
-    "@trpc/server": "^10.18.0",
-    "@upstash/ratelimit": "^0.4.2",
-    "@upstash/redis": "^1.20.4",
-    "class-variance-authority": "^0.6.0",
-    "clsx": "^1.2.1",
-    "cuid": "^3.0.0",
-    "drizzle-orm": "^0.25.3",
-    "en": "link:javascript-time-ago/locale/en",
-    "javascript-time-ago": "^2.5.9",
-    "lucide-react": "^0.187.0",
-    "next": "^13.2.4",
-    "next-auth": "^4.21.0",
+    "@auth/drizzle-adapter": "^0.3.2",
+    "@planetscale/database": "^1.11.0",
+    "@t3-oss/env-nextjs": "^0.7.0",
+    "@tanstack/react-query": "^4.32.6",
+    "@trpc/client": "^10.37.1",
+    "@trpc/next": "^10.37.1",
+    "@trpc/react-query": "^10.37.1",
+    "@trpc/server": "^10.37.1",
+    "drizzle-orm": "^0.28.5",
+    "next": "^13.5.4",
+    "next-auth": "^4.23.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "shadcn-ui": "0.1.1",
-    "superjson": "1.12.2",
-    "tailwind-merge": "^1.12.0",
-    "tailwindcss-animate": "^1.0.5",
-    "zod": "^3.21.4"
+    "superjson": "^1.13.1",
+    "zod": "^3.22.4"
   },
   "devDependencies": {
-    "@types/eslint": "^8.21.3",
-    "@types/node": "^18.15.5",
-    "@types/prettier": "^2.7.2",
-    "@types/react": "^18.0.28",
-    "@types/react-dom": "^18.0.11",
-    "@typescript-eslint/eslint-plugin": "^5.56.0",
-    "@typescript-eslint/parser": "^5.56.0",
+    "@types/eslint": "^8.44.2",
+    "@types/node": "^18.16.0",
+    "@types/react": "^18.2.20",
+    "@types/react-dom": "^18.2.7",
+    "@typescript-eslint/eslint-plugin": "^6.3.0",
+    "@typescript-eslint/parser": "^6.3.0",
     "autoprefixer": "^10.4.14",
-    "eslint": "^8.36.0",
-    "eslint-config-next": "^13.2.4",
-    "postcss": "^8.4.21",
-    "prettier": "^2.8.6",
-    "prettier-plugin-tailwindcss": "^0.2.6",
-    "prisma": "^4.11.0",
-    "tailwindcss": "^3.3.0",
-    "typescript": "^5.0.2"
+    "dotenv-cli": "^7.3.0",
+    "drizzle-kit": "^0.19.13",
+    "eslint": "^8.47.0",
+    "eslint-config-next": "^13.5.4",
+    "mysql2": "^3.6.1",
+    "postcss": "^8.4.27",
+    "prettier": "^3.0.0",
+    "prettier-plugin-tailwindcss": "^0.5.1",
+    "tailwindcss": "^3.3.3",
+    "typescript": "^5.1.6"
   },
   "ct3aMetadata": {
-    "initVersion": "7.12.1"
+    "initVersion": "7.22.0"
   }
 }
