@@ -1,6 +1,5 @@
-import { createTRPCRouter } from "@/server/api/trpc";
-import { exampleRouter } from "@/server/api/routers/example";
-import { linkRouter } from "@/server/api/routers/link";
+import { postRouter } from "~/server/api/routers/post";
+import { createTRPCRouter } from "~/server/api/trpc";
 
 /**
  * This is the primary router for your server.
@@ -8,8 +7,7 @@ import { linkRouter } from "@/server/api/routers/link";
  * All routers added in /api/routers should be manually added here.
  */
 export const appRouter = createTRPCRouter({
-  example: exampleRouter,
-  link: linkRouter,
+  post: postRouter,
 });
 
 // export type definition of API
